+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Added
 
 -   Support passing extra arguments to the runner
+-   Added a common language section for the runner path
 
 ## [5.0.0] - 2024-07-12